Make kvmd compatible with libgpiod 1.2.
The singular default_val argument of gpiod.Line.request() method
was introduced in libgpiod 1.3.

For older versions of libgpiod, defailt_vals argument with list value
should be used.

This argument is available in newer versions of libgpiod as well
for compatibility.

This change is needed for Debian / Raspbian 10 that have libgpiod 1.2.

Signed-off-by: Oleg Girko <ol@infoserver.lv>