Fix crash on unload (#136)
* Fix crash on unload

Within main, exists a `window.beforeunload` handler which brings up the "Are you sure you want to close PiKVM session?" message.   When the page is refreshed, the event is `undefined` and, the code which sets the `event.returnValue` to the aforementioned text provides an exception.  

To reproduce:
1. Open KVMD web
2. Open developer tools to view Console.
3. Interact with the page by clicking the main window.
4. Refresh the page
5. Observe type error in console. 

This patch checks if the event is defined before attempting to set the `event.returnValue`.  Other functions are maintained.

* Add space